BasicMemberDiscount class created for members that do not have a discount rate for any books

+++ src/Members/Library/member.java
@@ -1,5 +1,5 @@
 package Members.Library;
 
 public interface member {
-    public double memberFee();
+    public int memberFee();
 }
